+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/Views/MainWindow.xaml.cs
@@ -10,18 +10,33 @@ namespace TD.WPF.Views
     {
         private readonly AppSettings _appSettings = AppSettings.GetAppSettings();
         private readonly ITradeRepository _tradeRepository;
+        private readonly ICurrencyRepository _currencyRepository;
+        private readonly IExchangeRepository _exchangeRepository;
+        private readonly ISymbolRepository _symbolRepository;
+        private readonly ITradingAccountRepository _tradingAccountRepository;
         private MainWindowVM _viewModel;
 
         /// <summary>
         /// Parameterless constructor for the XAML designer.
         /// </summary>
-        public MainWindow() : this(new TradeMockRepository())
+        public MainWindow() : this(
+            new TradeMockRepository(),
+            new CurrencyMockRepository(), new ExchangeMockRepository(),
+            new SymbolMockRepository(), new TradingAccountMockRepository())
         {
         }
 
-        public MainWindow(ITradeRepository tradeRepository)
+        public MainWindow(ITradeRepository tradeRepository,
+                          ICurrencyRepository currencyRepository,
+                          IExchangeRepository exchangeRepository,
+                          ISymbolRepository symbolRepository,
+                          ITradingAccountRepository tradingAccountRepository)
         {
             _tradeRepository = tradeRepository ?? throw new ArgumentNullException(nameof(tradeRepository));
+            _currencyRepository = currencyRepository ?? throw new ArgumentNullException(nameof(currencyRepository));
+            _exchangeRepository = exchangeRepository ?? throw new ArgumentNullException(nameof(exchangeRepository));
+            _symbolRepository = symbolRepository ?? throw new ArgumentNullException(nameof(symbolRepository));
+            _tradingAccountRepository = tradingAccountRepository ?? throw new ArgumentNullException(nameof(tradingAccountRepository));
             InitializeComponent();
             SourceInitialized += MainWindow_SourceInitialized;
             Closing += MainWindow_Closing;
@@ -32,7 +47,9 @@ namespace TD.WPF.Views
             SourceInitialized -= MainWindow_SourceInitialized;
             _appSettings.MainWindowPosition.SetWindowPositions(this);
             Title = $"Ozz Trade Diary - v{AppVersion.Version}";
-            _viewModel = new MainWindowVM(_tradeRepository);
+            _viewModel = new MainWindowVM(_tradeRepository,
+                                          _currencyRepository, _exchangeRepository,
+                                          _symbolRepository, _tradingAccountRepository);
             DataContext = _viewModel;
         }